% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ApiConfigHandler}
\alias{ApiConfigHandler}
\title{ApiConfigHandler Object}
\usage{
ApiConfigHandler(login = NULL, url = NULL, securityLevel = NULL,
  authFailAction = NULL, script = NULL)
}
\arguments{
\item{login}{Level of login required to access this resource}

\item{url}{URL to serve the endpoint at}

\item{securityLevel}{Security (HTTPS) enforcement for this URL}

\item{authFailAction}{Action to take when users access resources that require authentication}

\item{script}{Path to the script from the application root directory}
}
\value{
ApiConfigHandler object
}
\description{
ApiConfigHandler Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Google Cloud Endpoints (https://cloud.google.com/appengine/docs/python/endpoints/) configuration for API handlers.
}
