% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ApiEndpointHandler}
\alias{ApiEndpointHandler}
\title{ApiEndpointHandler Object}
\usage{
ApiEndpointHandler(scriptPath = NULL)
}
\arguments{
\item{scriptPath}{Path to the script from the application root directory}
}
\value{
ApiEndpointHandler object
}
\description{
ApiEndpointHandler Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Uses Google Cloud Endpoints to handle requests.
}
