% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_functions.R
\name{apps.create}
\alias{apps.create}
\title{Creates an App Engine application for a Google Cloud Platform project. Required fields:id - The ID of the target Cloud Platform project.location - The region (https://cloud.google.com/appengine/docs/locations) where you want the App Engine application located.For more information about App Engine applications, see Managing Projects, Applications, and Billing (https://cloud.google.com/appengine/docs/python/console/).}
\usage{
apps.create(Application)
}
\arguments{
\item{Application}{The \link{Application} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/appengine/docs/admin-api/}{Google Documentation}

Other Application functions: \code{\link{Application}},
  \code{\link{apps.patch}}
}
