% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{BasicScaling}
\alias{BasicScaling}
\title{BasicScaling Object}
\usage{
BasicScaling(maxInstances = NULL, idleTimeout = NULL)
}
\arguments{
\item{maxInstances}{Maximum number of instances to create for this version}

\item{idleTimeout}{Duration of time after the last request that an instance must wait before the instance is shut down}
}
\value{
BasicScaling object
}
\description{
BasicScaling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A service with basic scaling will create an instance when the application receives a request. The instance will be turned down when the app becomes idle. Basic scaling is ideal for work that is intermittent or driven by user activity.
}
