% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ContainerInfo}
\alias{ContainerInfo}
\title{ContainerInfo Object}
\usage{
ContainerInfo(image = NULL)
}
\arguments{
\item{image}{URI to the hosted container image in Google Container Registry}
}
\value{
ContainerInfo object
}
\description{
ContainerInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Docker image that is used to create a container and start a VM instance for the version that you deploy. Only applicable for instances running in the App Engine flexible environment.
}
