% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Deployment.files}
\alias{Deployment.files}
\title{Deployment.files Object}
\usage{
Deployment.files()
}
\value{
Deployment.files object
}
\description{
Deployment.files Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Manifest of the files stored in Google Cloud Storage that are included as part of this version. All files must be readable using the credentials supplied with this call.
}
\seealso{
Other Deployment functions: \code{\link{Deployment}}
}
