% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{EndpointsApiService}
\alias{EndpointsApiService}
\title{EndpointsApiService Object}
\usage{
EndpointsApiService(name = NULL, configId = NULL)
}
\arguments{
\item{name}{Endpoints service name which is the name of the 'service' resource in the Service Management API}

\item{configId}{Endpoints service configuration id as specified by the Service Management API}
}
\value{
EndpointsApiService object
}
\description{
EndpointsApiService Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Endpoints (https://cloud.google.com/endpoints) configuration. The Endpoints API Service provides tooling for serving Open API and gRPC endpoints via an NGINX proxy.The fields here refer to the name and configuration id of a 'service' resource in the Service Management API (https://cloud.google.com/service-management/overview).
}
