% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Instance}
\alias{Instance}
\title{Instance Object}
\usage{
Instance(vmName = NULL, vmId = NULL, qps = NULL, name = NULL,
  vmZoneName = NULL, averageLatency = NULL, id = NULL,
  memoryUsage = NULL, vmIp = NULL, errors = NULL, vmStatus = NULL,
  availability = NULL, startTime = NULL, vmDebugEnabled = NULL,
  requests = NULL, appEngineRelease = NULL)
}
\arguments{
\item{vmName}{Name of the virtual machine where this instance lives}

\item{vmId}{Virtual machine ID of this instance}

\item{qps}{Average queries per second (QPS) over the last minute}

\item{name}{Full path to the Instance resource in the API}

\item{vmZoneName}{Zone where the virtual machine is located}

\item{averageLatency}{Average latency (ms) over the last minute}

\item{id}{Relative name of the instance within the version}

\item{memoryUsage}{Total memory in use (bytes)}

\item{vmIp}{The IP address of this instance}

\item{errors}{Number of errors since this instance was started}

\item{vmStatus}{Status of the virtual machine where this instance lives}

\item{availability}{Availability of the instance}

\item{startTime}{Time that this instance was started}

\item{vmDebugEnabled}{Whether this instance is in debug mode}

\item{requests}{Number of requests since this instance was started}

\item{appEngineRelease}{App Engine release this instance is running on}
}
\value{
Instance object
}
\description{
Instance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Instance resource is the computing unit that App Engine uses to automatically scale an application.
}
