% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ListVersionsResponse}
\alias{ListVersionsResponse}
\title{ListVersionsResponse Object}
\usage{
ListVersionsResponse(versions = NULL, nextPageToken = NULL)
}
\arguments{
\item{versions}{The versions belonging to the requested service}

\item{nextPageToken}{Continuation token for fetching the next page of results}
}
\value{
ListVersionsResponse object
}
\description{
ListVersionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for Versions.ListVersions.
}
