% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ReadinessCheck}
\alias{ReadinessCheck}
\title{ReadinessCheck Object}
\usage{
ReadinessCheck(unhealthyThreshold = NULL, path = NULL, host = NULL,
  healthyThreshold = NULL, checkInterval = NULL, timeout = NULL)
}
\arguments{
\item{unhealthyThreshold}{Number of consecutive failed checks required before removing traffic}

\item{path}{The request path}

\item{host}{Host header to send when performing a HTTP Readiness check}

\item{healthyThreshold}{Number of consecutive successful checks required before receiving traffic}

\item{checkInterval}{Interval between health checks}

\item{timeout}{Time before the check is considered failed}
}
\value{
ReadinessCheck object
}
\description{
ReadinessCheck Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Readiness checking configuration for VM instances. Unhealthy instances are removed from traffic rotation.
}
