% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{RequestUtilization}
\alias{RequestUtilization}
\title{RequestUtilization Object}
\usage{
RequestUtilization(targetRequestCountPerSecond = NULL,
  targetConcurrentRequests = NULL)
}
\arguments{
\item{targetRequestCountPerSecond}{Target requests per second}

\item{targetConcurrentRequests}{Target number of concurrent requests}
}
\value{
RequestUtilization object
}
\description{
RequestUtilization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Target scaling by request utilization. Only applicable for VM runtimes.
}
